Handle the deploy 🔥
