+++ .gitignore
@@ -65,3 +65,7 @@ node_modules/
 .env
 
 dist
+functions
+!functions/index.js
+!functions/package-lock.json
+!functions/package.json
