+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "dev": "webpack --watch --config=webpack.config.server.js",
-    "build": "webpack --config=webpack.config.server.js",
-    "deploy": "npm run build && firebase deploy"
+    "build:server": "webpack --config=webpack.config.server.js --env.production",
+    "build:client": "webpack --config=webpack.config.client.js --env.production",
+    "deploy": "npm run build:server && npm run build:client && firebase deploy"
   },
   "keywords": [],
   "author": "",
