+++ src/partials/foot.html
@@ -1,6 +1,6 @@
     <script>
         if ("serviceWorker" in navigator) {
-            navigator.serviceWorker.register("/service-worker.js");
+            navigator.serviceWorker.register("./service-worker.js");
         }
     </script>
     </body>
